[refactor(MainActivity.kt)] MainActivity app모듈로 이동
