--- core/data/booksearch/internal/src/main/java/com/kova700/internal/MyClass.kt
@@ -1,4 +1,0 @@
-package com.kova700.internal
-
-class MyClass {
-}
