+++ feature/search/src/main/kotlin/com/kova700/search/feature/detail/DetailViewModel.kt
@@ -3,7 +3,6 @@ package com.kova700.search.feature.detail
 import androidx.lifecycle.SavedStateHandle
 import androidx.lifecycle.ViewModel
 import com.kova700.amazonbookstorepractice.core.data.booksearch.external.usecase.GetCachedBookUseCase
-import com.kova700.search.feature.SELECTED_BOOK_INDEX
 import com.kova700.search.feature.mapper.toItem
 import com.kova700.search.feature.model.BookItem
 import dagger.hilt.android.lifecycle.HiltViewModel
@@ -12,7 +11,7 @@ import kotlinx.coroutines.flow.asStateFlow
 import javax.inject.Inject
 
 @HiltViewModel
-internal class DetailViewModel @Inject constructor(
+class DetailViewModel @Inject constructor(
 	private val savedStateHandle: SavedStateHandle,
 	private val getCachedBookUseCase: GetCachedBookUseCase
 ) : ViewModel() {
@@ -30,3 +29,4 @@ internal class DetailViewModel @Inject constructor(
 		_viewState.value = getCachedBookUseCase(index).toItem()
 	}
 }
+const val SELECTED_BOOK_INDEX = "SELECTED_BOOK_INDEX"
