+++ feature/search/src/main/kotlin/com/kova700/search/feature/detail/DetailWebView.kt
@@ -5,7 +5,7 @@ import androidx.compose.runtime.Composable
 import androidx.compose.ui.viewinterop.AndroidView
 
 @Composable
-internal fun DetailWebView(url: String) {
+fun DetailWebView(url: String) {
 	AndroidView(factory = {
 		WebView(it).apply {
 			loadUrl(url)
