+++ feature/search/src/main/kotlin/com/kova700/search/feature/search/SearchViewModel.kt
@@ -20,7 +20,7 @@ import kotlinx.coroutines.launch
 import javax.inject.Inject
 
 @HiltViewModel
-internal class SearchViewModel @Inject constructor(
+class SearchViewModel @Inject constructor(
 	savedStateHandle: SavedStateHandle,
 	private val getSearchedBookUseCase: GetSearchedBookUseCase,
 	private val getSearchedBookFlowUseCase: GetSearchedBookFlowUseCase,
